+++ CHANGELOG.rst
@@ -7,6 +7,22 @@ HEAD
 ----
 * Fix bug where fetching items from a folder that can contain multiple item types (e.g. the Deleted Items folder) would
   only return one item type.
+* ``Folder.bulk_update()`` was changed to expect a map with ``Item`` instance as key and a list of changed fieldnames on
+  each item as value. E.g.:
+
+  .. code-block:: python
+
+      items = []
+      for i in range(4):
+          item = Message(subject='Test %s' % i)
+          items.append(item)
+      account.sent.bulk_create(items=items)
+
+      item_changes = {}
+      for i, item in enumerate(items):
+          item.subject = 'Changed subject' % i
+          changed_items[item] = 'subject'
+      account.sent.bulk_update(items=item_changes)
 
 
 1.7.0